correct anthropic prompts
