+++ packages/server/customer-os-common-module/config/common_config.go
@@ -33,7 +33,6 @@ type InternalServicesConfig struct {
 
 type ExternalServicesConfig struct {
 	AnthropicConfig      AnthropicConfig
-	AnthropicPrompts     AnthropicPrompts
 	BetterContactConfig  BetterContactConfig
 	BrandfetchConfig     BrandfetchConfig
 	CloudflareConfig     CloudflareConfig
